requirements.txt
discord.py
aiohttp
websocket
requests
python-dateutil
beautifulsoup4
markdownify
html2text
google-api-python-client
oauth2client
